# pkg/swagger/models/k8s_resource.go
package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-swagger/go-swagger/strfmt"

	"github.com/go-swagger/go-swagger/errors"
)

/*K8sResource k8s resource

swagger:model k8sResource
*/
type K8sResource struct {

	/* data

	Required: true
	*/
	Data interface{} `json:"data"`
}

// Validate validates this k8s resource
func (m *K8sResource) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateData(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *K8sResource) validateData(formats strfmt.Registry) error {

	return nil
}